[vim] Remove <leader>c git commit mapping
I was not using it. The proof is that it now gives "Gcommit is
deprecated in favour of Git commit"

I prefer to use mappings inside of the status window. This allows me to
create leader mappings starting with c without any conflict. The only
issue would be that this mapping takes a few seconds to open,
which would not be the end of the world given that it's not used,
but I prefer to clean this up